% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spline.R
\name{build_spline}
\alias{build_spline}
\title{Generate univariate B-spline expansions}
\usage{
build_spline(X, df)
}
\arguments{
\item{X}{A $n x p$ matrix of covariates from which you want to evaluate
all marginal spline effects.}

\item{df}{The degree of freedom for the basis function expansion}
}
\value{
An array of univariate B-spline function expansions. The first dimension is
the number of variables $p$. The second dimension is the number
of observations $n$ and the third dimension is the number of knots in the
basis expansion $df$.
}
\description{
Generate univariate B-spline expansions
}
\examples{

library(splines)
set.seed(111)
X <- matrix(rnorm(40), 10, 4) ## 10 observations and 4 variables

Z <- build_spline(X, 5)

}
