% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spline_effects.R
\name{build_spline_effects}
\alias{build_spline_effects}
\title{Build the non-linear response from the basis functions and paramters}
\usage{
build_spline_effects(Z, beta)
}
\arguments{
\item{Z}{A q by n by df array of spline-expanded coefficients}

\item{beta}{A q by df matrix of spline coefficieints}
}
\value{
A n by q matrix of effects (non-linear functional response)
}
\description{
Build the non-linear response from the basis functions and paramters
}
