% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_linear_fixed.R
\name{lm_linear_fixed}
\alias{lm_linear_fixed}
\title{lm_linear_fixed}
\usage{
lm_linear_fixed(
  y,
  X,
  n_plot,
  n_tree,
  plot_by_tree_idx,
  tree_idx,
  X_pred,
  tree_idx_pred,
  ...
)
}
\arguments{
\item{y}{A vector of length n of tree ring increments on a log scale}

\item{X}{A n times p matrix of covariate values}

\item{n_plot}{An integer for the number of plots}

\item{n_tree}{An integer for the number of trees}

\item{plot_by_tree_idx}{A vector of length n_tree that indicates which plot each tree was sampled from}

\item{tree_idx}{A vector of length n indicating which tree the observation came from}

\item{X_pred}{A n_pred by p matrix of covariate values at which to generate predictions}

\item{tree_idx_pred}{A vector of length n_pred indicating which tree the observation came from}

\item{...}{Additional arguments to \code{rstan::sampling()}}
}
\value{
A stan object that is the result of fitting the linear regression model
}
\description{
lm_linear_fixed
}
