% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_linear_X.R
\name{lm_linear_X}
\alias{lm_linear_X}
\title{lm_linear_no_p}
\usage{
lm_linear_X(y, X, X_int, X_pred, X_int_pred, ...)
}
\arguments{
\item{y}{A vector of length n of tree ring increments on a log scale}

\item{X}{A n times p matrix of covariate values}

\item{X_int}{A n times n_tree matrix of covariate values}

\item{X_pred}{A n_pred by p matrix of covariate values at which to generate predictions}

\item{...}{Additional arguments to \code{rstan::sampling()}}
}
\value{
A stan object that is the result of fitting the linear regression model
}
\description{
lm_linear_no_p
}
