% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_splines.R
\name{lm_splines}
\alias{lm_splines}
\title{lm_splines}
\usage{
lm_splines(
  y,
  X,
  X_bs,
  n_plot,
  n_tree,
  plot_by_tree_idx,
  tree_idx,
  X_pred,
  X_bs_pred,
  tree_idx_pred,
  ...
)
}
\arguments{
\item{y}{A vector of length n of tree ring increments on a log scale}

\item{X}{A n times p matrix of covariate values}

\item{X_bs}{A p times n times df array of covariate values bspline basis function expansions}

\item{n_plot}{An integer for the number of plots}

\item{n_tree}{An integer for the number of trees}

\item{tree_idx}{A vector of length n indicating which tree the observation came from}

\item{X_pred}{A n_pred by p matrix of covariate values at which to generate predictions}

\item{X_bs_pred}{A p times n_pred times df array of covariate values bspline basis function expansions}

\item{tree_idx_pred}{A vector of length n_pred indicating which tree the observation came from}

\item{...}{Additional argument to `rstan::sampling()`}

\item{plot_idx}{A vector of length n_tree that indicates which plot each tree was sampled from}
}
\value{
A stan object that is the result of fitting the spline regression model
}
\description{
lm_splines
}
