% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Q.R
\name{make_Q}
\alias{make_Q}
\title{Generate spline penalty precision matrix}
\usage{
make_Q(df, phi, prec_model = "CAR")
}
\arguments{
\item{df}{is the degree of freedom parameter}

\item{phi}{is a number between -1 and 1 that defines the strength of the autoregressive process. Typically this will be set to 1 for use as a prior in penalized Bayesian models}

\item{prec_model}{is a string that takes the values "CAR" or "SAR" and defines the graphical structure for the precision matrix.}
}
\value{
a list of  \eqn{n \times n}{n x n} sparse spam matrices or Matrix matrices of class "dgCMatrix" (see Matrix package for details)
}
\description{
A function for setting up a conditional autoregressive (CAR) or simultaneous autoregressive (SAR) precision matrix for use as a prior in Bayesian models
}
\examples{
df <- 4
phi <- 0.8
Q <- make_Q(df, phi)

}
