% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_W.R
\name{make_W}
\alias{make_W}
\title{Generate spline penalty adjacency matrix}
\usage{
make_W(df)
}
\arguments{
\item{df}{is the degree of freedom parameter}
}
\value{
A \eqn{df \times df}{df x df} adjacency matrices
}
\description{
A function for setting up a conditional autoregressive (CAR) or simultaneous autoregressive (SAR) precision matrix for use as a prior in Bayesian models
}
\examples{
df <- 4
phi <- 0.8
Q <- make_Q(df, phi)

}
