% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_splines.R
\name{mcmc_splines}
\alias{mcmc_splines}
\title{lm_linear}
\usage{
mcmc_splines(
  y,
  X,
  Xbs,
  plot_idx,
  tree_idx,
  params,
  config = NULL,
  verbose = FALSE,
  n_chain = 1
)
}
\arguments{
\item{y}{A vector of length n of tree ring increments on a log scale}

\item{X}{A n times p matrix of covariate values}

\item{Xbs}{A n times df*q matrix of b-spline expanded covariate values}

\item{plot_idx}{A vector of length n that indicates which plot each observation was sampled from}

\item{tree_idx}{A vector of length n indicating which tree the observation came from}

\item{params}{is a list of parameter settings. The list
\code{params} must contain the following values:
\itemize{
\item \code{n_adapt}: A positive integer number of adaptive MCMC iterations.
\item \code{n_mcmc}: A positive integer number of total MCMC iterations
post adaptation.
\item \code{n_thin}: A positive integer number of MCMC iterations per saved
sample.
\item \code{n_message}: A positive integer number of frequency of iterations
to output a progress message. For example, \code{n_message = 50}
outputs progress messages every 50 iterations.
}}

\item{config}{is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.}

\item{verbose}{Should verbose output be printed? Typically this is only useful for troubleshooting.}

\item{n_chain}{is the MCMC chain id. The default is 1.}

\item{df}{The b-spline degree of freedom parameter}
}
\value{
A MCMC object that is the result of fitting the linear regression model
}
\description{
lm_linear
}
