% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_interactions_lattice.R
\name{spline_interactions_lattice}
\alias{spline_interactions_lattice}
\title{Generate separable B-spline interactions multivariate regression penalty
precision matrix for prior}
\usage{
spline_interactions_lattice(df)
}
\arguments{
\item{df}{The degree of freedom for the basis function expansion}
}
\value{
A matrix with df^2 rows and 2 columns that indicate which basis
    functions are used in the interaction
}
\description{
Generate separable B-spline interactions multivariate regression penalty
precision matrix for prior
}
\examples{

set.seed(111)
df <- 5

Q <- spline_interactions_penalty(5)

}
