% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_interactions_penalty.R
\name{spline_interactions_penalty}
\alias{spline_interactions_penalty}
\title{Generate separable B-spline interactions multivariate regression penalty
precision matrix for prior}
\usage{
spline_interactions_penalty(df, phi = 1)
}
\arguments{
\item{df}{The degree of freedom for the basis function expansion}

\item{phi}{is a number between -1 and 1 that defines the strength of the autoregressive process. Typically this will be set to 1 for use as a prior in penalized Bayesian models}
}
\value{
A precision matrix for the smoothing penalty prior
}
\description{
Generate separable B-spline interactions multivariate regression penalty
precision matrix for prior
}
\examples{

set.seed(111)
df <- 5

Q <- spline_interactions_penalty(5)

}
