% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeRingSplines-package.R
\docType{package}
\name{treeRingSplines-package}
\alias{treeRingSplines-package}
\alias{treeRingSplines}
\title{The 'treeRingSplines' package.}
\description{
Fitting Tree Ring Growth Models With Splines.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
